requirements.txt
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
arrow==1.0.3
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bank-card==1.0.5
base58==2.1.0
beautifulsoup4==4.8.0
bitarray==1.2.2
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cbpro==1.1.4
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
coloredlogs==15.0
conda==4.8.3
conda-build==3.18.9
conda-package-handling==1.7.0
conda-verify==3.4.2
configobj==5.0.6
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.11.0
dask==2.5.2
dateparser==1.0.0
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.1
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Faker==4.14.0
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.2
future==0.18.2
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.9.0
HeapDict==1.0.1
hexbytes==0.2.1
html5lib==1.0.1
humanfriendly==9.1
hyperlink==21.0.0
idna==2.9
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
incremental==21.3.0
ipfshttpclient==0.7.0a1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
loguru==0.5.3
lru-dict==1.1.7
lxml==4.5.1
macholib==1.14
MarkupSafe==1.1.1
marshmallow==3.11.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
moment==0.12.1
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multiaddr==0.0.9
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
netaddr==0.8.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
numpydoc==0.9.1
olefile==0.46
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pipenv==2020.11.15
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.17.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
py-etherscan-api==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.10.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pylint==2.4.2
pymongo==3.5.1
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-binance==0.7.11
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
rlp==2.0.1
rope==0.14.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
timeout-decorator==0.5.0
times==0.7
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
unicodecsv==0.14.1
urllib3==1.25.9
varint==1.0.2
virtualenv==20.4.4
virtualenv-clone==0.5.4
wcwidth==0.1.7
web3==5.19.0
webencodings==0.5.1
websocket-client==0.40.0
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
xmltodict==0.12.0
zict==1.0.0
zipp==0.6.0
zope.interface==5.4.0
